Updated and redifined Login endpoints and related user models.
